Add keybinds for temporarily releasing ui control lock Make urls clickable in about modal Fix push/pop id mismatch in actor info window

+++ ImmersiveEquipmentDisplays/Drivers/Input/KeyEventType.h
@@ -2,14 +2,14 @@
 
 namespace IED
 {
-	enum class KeyEventType : std::uint32_t
+	enum class KeyEventType : std::uint8_t
 	{
 		Keyboard = 0,
 		Mouse    = 1,
 		Wheel    = 2
 	};
 
-	enum class KeyEventState : std::uint32_t
+	enum class KeyEventState : std::uint8_t
 	{
 		KeyDown = 0,
 		KeyUp   = 1
